adjusted Activity personal list , added stock
